move properties from Mesh_3 to BGL, Polyhedron_3 and Surface_mesh
