+++ Mesh_3/examples/Mesh_3/mesh_polyhedral_domain_sm.cpp
@@ -8,7 +8,7 @@
 #include <CGAL/Polyhedral_mesh_domain_3.h>
 #include <CGAL/make_mesh_3.h>
 #include <CGAL/refine_mesh_3.h>
-#include <CGAL/Mesh_3/properties_Surface_mesh.h>
+
 // Domain
 typedef CGAL::Exact_predicates_inexact_constructions_kernel K;
 typedef CGAL::Surface_mesh<K::Point_3> Polyhedron;
