+++ Mesh_3/include/CGAL/Polyhedral_mesh_domain_with_features_3.h
@@ -42,7 +42,6 @@
 #include <CGAL/Mesh_3/Detect_polylines_in_polyhedra.h>
 #include <CGAL/Mesh_3/Polyline_with_context.h>
 #include <CGAL/Polygon_mesh_processing/detect_features.h>
-#include <CGAL/Mesh_3/properties_Polyhedron_3.h>
 
 #include <CGAL/enum.h>
 #include <CGAL/IO/Polyhedron_iostream.h>
