+++ Polygon_mesh_processing/examples/Polygon_mesh_processing/detect_features_example.cpp
@@ -1,8 +1,6 @@
 #include <CGAL/Exact_predicates_inexact_constructions_kernel.h>
 #include <CGAL/Surface_mesh.h>
-
 #include <CGAL/Polygon_mesh_processing/detect_features.h>
-#include <CGAL/Mesh_3/properties_Surface_mesh.h>
 
 #include <fstream>
