+++ Polygon_mesh_processing/test/Polygon_mesh_processing/test_detect_features.cpp
@@ -1,6 +1,5 @@
 #include <CGAL/Exact_predicates_inexact_constructions_kernel.h>
 #include <CGAL/Surface_mesh.h>
-#include <CGAL/Mesh_3/properties_Surface_mesh.h>
 
 #include <CGAL/Polygon_mesh_processing/detect_features.h>
