+++ Surface_mesh/include/CGAL/boost/graph/properties_Surface_mesh_features.h
@@ -18,11 +18,10 @@
 //
 // Author(s)     : Andreas Fabri
 
-#ifndef CGAL_MESH_3_PROPERTIES_SURFACE_MESH_H
-#define CGAL_MESH_3_PROPERTIES_SURFACE_MESH_H
+#ifndef CGAL_PROPERTIES_SURFACE_MESH_FEATURES_H
+#define CGAL_PROPERTIES_SURFACE_MESH_FEATURES_H
 
 #include <CGAL/Surface_mesh.h>
-#include <CGAL/Mesh_3/properties.h>
 #include <set>
 
 namespace boost {
@@ -81,46 +80,9 @@ struct property_map<CGAL::Surface_mesh<P>, CGAL::vertex_incident_patches_t<I> >
 };
 
 
-template <typename P>
-struct property_map<CGAL::Surface_mesh<P>, CGAL::vertex_time_stamp_t>
-{
-
-  typedef typename boost::graph_traits<CGAL::Surface_mesh<P> >::vertex_descriptor vertex_descriptor;
-
-  typedef typename CGAL::Surface_mesh<P>::template Property_map<vertex_descriptor, std::size_t> type;
-  typedef type const_type;
-};
-
-
-template <typename P>
-struct property_map<CGAL::Surface_mesh<P>, CGAL::halfedge_time_stamp_t>
-{
-
-  typedef typename boost::graph_traits<CGAL::Surface_mesh<P> >::halfedge_descriptor halfedge_descriptor;
-
-  typedef typename CGAL::Surface_mesh<P>::template Property_map<halfedge_descriptor, std::size_t> type;
-  typedef type const_type;
-};
-
-
-template <typename P>
-struct property_map<CGAL::Surface_mesh<P>, CGAL::face_time_stamp_t>
-{
-
-  typedef typename boost::graph_traits<CGAL::Surface_mesh<P> >::face_descriptor face_descriptor;
-
-  typedef typename CGAL::Surface_mesh<P>::template Property_map<face_descriptor, std::size_t> type;
-  typedef type const_type;
-};
-
 } // namespace boost
 
 namespace CGAL {
-
-template <typename P, typename Property_tag>
-struct Get_pmap_of_surface_mesh {
-  typedef typename boost::property_map<Surface_mesh<P>, Property_tag >::type type;
-};
 
 template <typename P, typename I>
 typename boost::lazy_disable_if<
@@ -146,14 +108,14 @@ inline get(CGAL::face_patch_id_t<I>, const Surface_mesh<P> & smesh)
 }
 
 
-#define CGAL_MESH_3_PROPERTY_SURFACE_MESH_RETURN_TYPE(Tag) \
+#define CGAL_PROPERTY_SURFACE_MESH_RETURN_TYPE(Tag) \
   typename boost::lazy_disable_if<                      \
      boost::is_const<P>,                                \
      Get_pmap_of_surface_mesh<P, Tag >                  \
    >::type
 
 template <typename P>
-CGAL_MESH_3_PROPERTY_SURFACE_MESH_RETURN_TYPE(CGAL::face_patch_id_t<void>)
+CGAL_PROPERTY_SURFACE_MESH_RETURN_TYPE(CGAL::face_patch_id_t<void>)
 inline get(CGAL::face_patch_id_t<void>, const Surface_mesh<P> &)
 {
   typedef CGAL::Static_property_map<typename boost::graph_traits<Surface_mesh<P> >::face_descriptor,std::pair<int,int> > Pmap;
@@ -163,7 +125,7 @@ inline get(CGAL::face_patch_id_t<void>, const Surface_mesh<P> &)
 
 
 template <typename P>
-CGAL_MESH_3_PROPERTY_SURFACE_MESH_RETURN_TYPE(CGAL::edge_is_feature_t)
+CGAL_PROPERTY_SURFACE_MESH_RETURN_TYPE(CGAL::edge_is_feature_t)
 inline get(CGAL::edge_is_feature_t, Surface_mesh<P>& smesh)
 {
   typedef typename boost::graph_traits<Surface_mesh<P> >::edge_descriptor edge_descriptor;
@@ -172,7 +134,7 @@ inline get(CGAL::edge_is_feature_t, Surface_mesh<P>& smesh)
 
 
 template <typename P>
-CGAL_MESH_3_PROPERTY_SURFACE_MESH_RETURN_TYPE(CGAL::edge_is_feature_t)
+CGAL_PROPERTY_SURFACE_MESH_RETURN_TYPE(CGAL::edge_is_feature_t)
 inline get(CGAL::edge_is_feature_t, const Surface_mesh<P>& smesh)
 {
   typedef typename boost::graph_traits<Surface_mesh<P> >::edge_descriptor edge_descriptor;
@@ -181,7 +143,7 @@ inline get(CGAL::edge_is_feature_t, const Surface_mesh<P>& smesh)
 
 
 template <typename P>
-CGAL_MESH_3_PROPERTY_SURFACE_MESH_RETURN_TYPE(CGAL::vertex_feature_degree_t)
+CGAL_PROPERTY_SURFACE_MESH_RETURN_TYPE(CGAL::vertex_feature_degree_t)
 inline get(CGAL::vertex_feature_degree_t, Surface_mesh<P> & smesh)
 {
   typedef typename boost::graph_traits<Surface_mesh<P> >::vertex_descriptor vertex_descriptor;
@@ -189,7 +151,7 @@ inline get(CGAL::vertex_feature_degree_t, Surface_mesh<P> & smesh)
 }
 
 template <typename P>
-CGAL_MESH_3_PROPERTY_SURFACE_MESH_RETURN_TYPE(CGAL::vertex_feature_degree_t)
+CGAL_PROPERTY_SURFACE_MESH_RETURN_TYPE(CGAL::vertex_feature_degree_t)
 inline get(CGAL::vertex_feature_degree_t, const Surface_mesh<P> & smesh)
 {
   typedef typename boost::graph_traits<Surface_mesh<P> >::vertex_descriptor vertex_descriptor;
@@ -218,55 +180,8 @@ typename boost::lazy_disable_if<
   return smesh. template property_map<vertex_descriptor,std::set<I> >("v:ip").first;
 }
 
-template <typename P>
-CGAL_MESH_3_PROPERTY_SURFACE_MESH_RETURN_TYPE(CGAL::vertex_time_stamp_t)
-inline get(CGAL::vertex_time_stamp_t, Surface_mesh<P> & smesh)
-{
-  typedef typename boost::graph_traits<Surface_mesh<P> >::vertex_descriptor vertex_descriptor;
-  return smesh. template add_property_map<vertex_descriptor,std::size_t>("v:time_stamp").first;
-}
-
-template <typename P>
-CGAL_MESH_3_PROPERTY_SURFACE_MESH_RETURN_TYPE(CGAL::vertex_time_stamp_t)
-inline get(CGAL::vertex_time_stamp_t, const Surface_mesh<P> & smesh)
-{
-  typedef typename boost::graph_traits<Surface_mesh<P> >::vertex_descriptor vertex_descriptor;
-  return smesh. template property_map<vertex_descriptor,std::size_t>("v:time_stamp").first;
-}
-
-template <typename P>
-CGAL_MESH_3_PROPERTY_SURFACE_MESH_RETURN_TYPE(CGAL::halfedge_time_stamp_t)
-inline get(CGAL::halfedge_time_stamp_t, Surface_mesh<P> & smesh)
-{
-  typedef typename boost::graph_traits<Surface_mesh<P> >::halfedge_descriptor halfedge_descriptor;
-  return smesh. template add_property_map<halfedge_descriptor,std::size_t>("h:time_stamp").first;
-}
-
-template <typename P>
-CGAL_MESH_3_PROPERTY_SURFACE_MESH_RETURN_TYPE(CGAL::halfedge_time_stamp_t)
-inline get(CGAL::halfedge_time_stamp_t, const Surface_mesh<P> & smesh)
-{
-  typedef typename boost::graph_traits<Surface_mesh<P> >::halfedge_descriptor halfedge_descriptor;
-  return smesh. template property_map<halfedge_descriptor,std::size_t>("h:time_stamp").first;
-}
-
-template <typename P>
-CGAL_MESH_3_PROPERTY_SURFACE_MESH_RETURN_TYPE(CGAL::face_time_stamp_t)
-inline get(CGAL::face_time_stamp_t, Surface_mesh<P> & smesh)
-{
-  typedef typename boost::graph_traits<Surface_mesh<P> >::face_descriptor face_descriptor;
-  return smesh. template add_property_map<face_descriptor,std::size_t>("v:time_stamp").first;
-}
-
-template <typename P>
-CGAL_MESH_3_PROPERTY_SURFACE_MESH_RETURN_TYPE(CGAL::face_time_stamp_t)
-inline get(CGAL::face_time_stamp_t, const Surface_mesh<P> & smesh)
-{
-  typedef typename boost::graph_traits<Surface_mesh<P> >::face_descriptor face_descriptor;
-  return smesh. template property_map<face_descriptor,std::size_t>("v:time_stamp").first;
-}
 } // namespace CGAL
 
-#undef CGAL_MESH_3_PROPERTY_SURFACE_MESH_RETURN_TYPE
+#undef CGAL_PROPERTY_SURFACE_MESH_RETURN_TYPE
 
-#endif //CGAL_MESH_3_PROPERTIES_SURFACE_MESH_H
+#endif //CGAL_PROPERTIES_SURFACE_MESH_FEATURES_H
